Script which replaces demo.sh for running Parsey McParseface model
This new script replaces demo.sh for running Parsey McParseface model.
It takes a different approach from demo.sh which is that it builds the
entire operation chain (tagging AND parsing) in a single model, complete
with a document source and document sink. The script also exports the model
in various ways, including using the TF Serving exporter, such that the
model can be served using TF Serving